Revert type changes

+++ src/sockets/BaseSocket.cpp
@@ -45,7 +45,7 @@ namespace sockets {
 		_recvTimeout = recvTimeout;
 
 #if defined(_WIN32) || defined(WIN32) || defined(_WIN64)
-                uint32_t ms = (uint32_t)recvTimeout.count();
+                uint32_t ms = recvTimeout.count();
                 setsockopt(_sockedFd, SOL_SOCKET, SO_RCVTIMEO, (const char*)&ms, sizeof ms);
 #else
                 timeval tv = makePortableInterval(recvTimeout);
